WIP: Update MRI upstreams for Phosphorus
Adopt latest versions, namely;
- odlparent-9.0.2
- yangtools-7.0.2-SNAPSHOT
- infrautils-2.0.2
- mdsal-8.0.0-SNAPSHOT
- controller-4.0.0-SNAPSHOT
- aaa-0.14.0-SNAPSHOT
- netconf-2.0.0-SNAPSHOT

Also introduce model-{inventory,topology}, which were removed from
upstream.

Since these versions also force the Uint8/Uint32 vs. Short/Long
equivalence issue, we mass-migrate all APIs to use Uint contructucts.

+++ extension/openflowplugin-extension-api/src/main/java/org/opendaylight/openflowplugin/extension/api/ExtensionConvertorData.java
@@ -10,6 +10,7 @@ package org.opendaylight.openflowplugin.extension.api;
 
 import org.opendaylight.yangtools.yang.common.Uint32;
 import org.opendaylight.yangtools.yang.common.Uint64;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 public class ExtensionConvertorData extends ConvertorData {
     private Uint32 xid;
@@ -22,7 +23,7 @@ public class ExtensionConvertorData extends ConvertorData {
      *
      * @param version the version
      */
-    public ExtensionConvertorData(final short version) {
+    public ExtensionConvertorData(final Uint8 version) {
         super(version);
     }
