+++ openflowplugin/src/main/java/org/opendaylight/openflowplugin/openflow/md/core/sal/convertor/action/data/ActionConvertorData.java
@@ -5,7 +5,6 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at http://www.eclipse.org/legal/epl-v10.html
  */
-
 package org.opendaylight.openflowplugin.openflow.md.core.sal.convertor.action.data;
 
 import org.opendaylight.openflowplugin.openflow.md.core.sal.convertor.data.VersionDatapathIdConvertorData;
@@ -23,7 +22,7 @@ public class ActionConvertorData extends VersionDatapathIdConvertorData {
      *
      * @param version the version
      */
-    public ActionConvertorData(final short version) {
+    public ActionConvertorData(final Uint8 version) {
         super(version);
     }
